+++ google-cloud/container-linux/kubernetes/workers/variables.tf
@@ -36,7 +36,7 @@ variable "machine_type" {
 
 variable "os_image" {
   type        = string
-  description = "Container Linux image for compute instanges (e.g. gcloud compute images list)"
+  description = "Flatcar Linux image for compute instanges (e.g. gcloud compute images list)"
 }
 
 variable "disk_size" {